move engine tests to a separate file
